chore: 엔티티 분리 - 각각의 패키지로 분리 - 불필요한 테스트 검증 코드 삭제
