+++ src/main/java/com/management/library/domain/management/ManagementRequestResult.java
@@ -1,5 +1,8 @@
-package com.management.library.domain;
+package com.management.library.domain.management;
 
+import com.management.library.domain.BaseEntity;
+import com.management.library.domain.admin.Administrator;
+import com.management.library.domain.management.ManagementRequest;
 import com.management.library.domain.type.RequestStatus;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -8,6 +11,7 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -22,18 +26,17 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EntityListeners(AuditingEntityListener.class)
-public class NewBookRequestResult extends BaseEntity {
+public class ManagementRequestResult extends BaseEntity {
 
-  @Id
-  @GeneratedValue
-  @Column(name = "new_book_request_result_id")
+  @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "management_reqeust_result_id")
   private Long id;
 
   @OneToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "new_book_request_id")
-  private NewBookRequest newBookRequest;
+  @JoinColumn(name = "management_request_id")
+  private ManagementRequest managementRequest;
 
-  @ManyToOne(fetch = FetchType.LAZY)
+  @ManyToOne
   @JoinColumn(name = "administrator_id")
   private Administrator administrator;
 
@@ -44,10 +47,11 @@ public class NewBookRequestResult extends BaseEntity {
   private RequestStatus result;
 
   @Builder
-  public NewBookRequestResult(Long id, NewBookRequest newBookRequest, Administrator administrator,
-      String resultPostTitle, String resultPostContent, RequestStatus result) {
+  public ManagementRequestResult(Long id, ManagementRequest managementRequest,
+      Administrator administrator, String resultPostTitle, String resultPostContent,
+      RequestStatus result) {
     this.id = id;
-    this.newBookRequest = newBookRequest;
+    this.managementRequest = managementRequest;
     this.administrator = administrator;
     this.resultPostTitle = resultPostTitle;
     this.resultPostContent = resultPostContent;
